Move the libanki.Utils.syncPlatform() method to AnkiDroidApp
Moved in the only class where it was used. It also didn't make sense
to have it in libanki.
